generator/requirements.txt
google-cloud-pubsub
google-cloud-storage
ratelimit==1.4.1